aura_agent_workshop_202510/pyproject.toml
[project]
name = "AuraAgentWorkshopDemo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
]
